Cleanup of pickler module
