+++ polymath/extensions/shrinker.py
@@ -3,8 +3,8 @@
 ################################################################################
 
 import numpy as np
-from ..qube   import Qube
-from ..scalar import Scalar
+from polymath.qube   import Qube
+from polymath.scalar import Scalar
 
 def shrink(self, antimask):
     """A 1-D version of this object, containing only the samples in the antimask
